requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
idna==3.7
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
